ClassAssignments/Class_04_10072025/Ristay_Wali_Aunty/pyproject.toml
[project]
name = "ristay-wali-aunty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.5.5",
    "chainlit>=2.6.2",
    "openai-agents>=0.2.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
